+++ noodles-bcf/src/writer.rs
@@ -126,7 +126,7 @@ where
     /// let mut writer = bcf::Writer::new(Vec::new());
     ///
     /// let header = vcf::Header::builder()
-    ///     .add_contig(Contig::new("sq0"))
+    ///     .add_contig(Contig::new("sq0".parse()?))
     ///     .build();
     ///
     /// writer.write_header(&header)?;